requirements.txt
transformers>=4.36.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
# Install torch separately if needed for your platform.
